backend/python-fastapi/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.7.4
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
reportlab>=4.0.0
yt-dlp>=2024.0.0
google-generativeai>=0.3.0

# LangChain
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-core>=0.3.0

# YouTube Search (SerpAPI)
httpx>=0.27.0
cachetools>=5.3.0
